pyproject.toml
[tool.poetry]
name = "traefik-certificate-exporter"
version = "0.0.10.0"
description = "Watches for changes to traefik acme json files and extracts certificates to a specific folder"
authors = ["Shawn Anderson <sanderson@eye-catcher.com>"]
readme = "README.md"
repository = "https://github.com/ravensorb/traefik-certificate-exporter"
homepage = "https://github.com/ravensorb/traefik-certificate-exporter"
keywords = ['traefik', 'letsencrypt']
packages = [{include = "traefik_certificate_exporter", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
watchdog = "^2.1.9"
docker = "^6.0.0"
requests = "^2.28.1"
docker-pycreds = "^0.4.0"
confuse = "^2.0.1"
poetry = "^1.8.2"
python-dotenv = "^1.0.1"
importlib-resources = "^6.4.0"
jsonpickle = "^3.0.3"
coloredlogs = "^15.0.1"
logging-tree = "^1.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
scriv = {extras = ["toml"], version = "^1.2.1"}

[tool.poetry.scripts]
traefik-certificate-exporter = "traefik_certificate_exporter.app:main"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = false
vcs = "git"
#style = "semver"
#format = "{base}.{distance}-{commit}"
format-jinja = """
    {%- if distance == 0 -%}
        {{ serialize_pep440(base, stage, revision) }}
    {%- elif revision is not none -%}
        {{ serialize_pep440(base, stage, revision, dev=distance) }}
    {%- else -%}
        {{ serialize_pep440(bump_version(base), stage, revision, dev=distance) }}
    {%- endif -%}
"""
bump = false

[tool.poetry-dynamic-versioning.substitution]
folders = [
  { path = "src" }
]
